feat: Inclusão do modal de exclusao do colaborador

+++ src/app/componentes/colaboradores/colaboradores.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { Location } from '@angular/common';
+import { CommonModule, Location } from '@angular/common';
 import { AppMarterialModule } from '../../compartilhado/app-material/app-material.module'; 
 import { Colaborador } from '../../models/colaborador.model';
 import { ColaboradoresService } from '../../services/colaboradores.service';
@@ -10,7 +9,7 @@ import { ErrorDialogComponent } from '../../compartilhado/components/error-dialo
 import { ActivatedRoute, Router } from '@angular/router';
 import { CompartilhadoListComponent } from '../../compartilhado/compartilhado-list/compartilhado-list.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
-
+import { ConfirmationexclusionDialogComponent } from '../../compartilhado/components/confirmationexclusion-dialog/confirmationexclusion-dialog.component';
 
 @Component({
   selector: 'app-colaboradores',
@@ -29,14 +28,12 @@ export class ColaboradoresComponent implements OnInit {
 
   constructor(
     private colaboradoresService: ColaboradoresService,
-    public  dialog: MatDialog,
+    public dialog: MatDialog,
     private router: Router,
     private route: ActivatedRoute,
     private location: Location,
-    private snackBar: MatSnackBar,
-  
+    private snackBar: MatSnackBar
   ) {
-
     this.colaboradores$ = this.colaboradoresService.list().pipe(
       catchError(error => {
         console.error('O Java deu erro! Detalhes:', error);
@@ -50,51 +47,59 @@ export class ColaboradoresComponent implements OnInit {
     );
   }
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
-  incluir() {
+  incluir(): void {
     this.router.navigate(['new'], { relativeTo: this.route });
   }
-  
-  edit(colaborador: Colaborador) {
-     this.router.navigate(['edit', colaborador.id], { relativeTo: this.route });
+
+  edit(colaborador: Colaborador): void {
+    this.router.navigate(['edit', colaborador.id], { relativeTo: this.route });
   }
 
-  refresh() {
+  refresh(): void {
     this.colaboradores$ = this.colaboradoresService.list();
   }
 
-  delete(colaborador: Colaborador) {
-  console.log('Excluindo o colaborador:', colaborador);
-  
-  this.colaboradoresService.delete(colaborador.id!).subscribe({
-    next: () => {
-      this.refresh();
-      this.snackBar.open('Colaborador excluído com sucesso!', 'Fechar', {
-        duration: 2500,
-        verticalPosition: 'top',
-        horizontalPosition: 'center'
-      });
-    },
-    error: (err) => {
-      console.error('Erro ao excluir:', err);
-      this.snackBar.open('Erro ao excluir colaborador.', 'Fechar', {
-        duration: 2500,
-        verticalPosition: 'top',
-        horizontalPosition: 'center'
-      });
-    }
-  });
-}
+  delete(colaborador: Colaborador): void {
+    console.log('Excluindo o colaborador:', colaborador);
 
-  openError(errorMsg: string) {
+    const dialogRef = this.dialog.open(ConfirmationexclusionDialogComponent, {
+      data: `Tem certeza que deseja excluir o colaborador "${colaborador.nomeColaborador}"?`,
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.colaboradoresService.delete(colaborador.id!).subscribe({
+          next: () => {
+            this.refresh();
+            this.snackBar.open('Colaborador excluído com sucesso!', 'Fechar', {
+              duration: 2500,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          },
+          error: (err) => {
+            console.error('Erro ao excluir:', err);
+            this.snackBar.open('Erro ao excluir colaborador.', 'Fechar', {
+              duration: 2500,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          }
+        });
+      }
+    });
+  }
+
+  openError(errorMsg: string): void {
     this.dialog.open(ErrorDialogComponent, {
       data: errorMsg
     });
   }
 
-  carregarColaboradores() {
+  carregarColaboradores(): void {
     console.log("Buscando dados no Java...");
   }
+
 }
